+++ packages/system-access/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/system-access
 
+## 1.0.14
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/axios-extension 2.0.10 → 2.1.0
+
+## 1.0.13
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/axios-extension 2.0.9 → 2.0.10
+
 ## 1.0.12
 
 ### Patch Changes